updated backend with updated requiremnts

+++ zakat-management-system/src/main/java/com/example/zakat/management/system/dtos/response/ReportResponse.java
@@ -1,13 +1,12 @@
 package com.example.zakat.management.system.dtos.response;
 
 import lombok.Data;
-import java.math.BigDecimal;
 
 @Data
 public class ReportResponse {
-    private BigDecimal totalDonated;
-    private BigDecimal totalDistributed;
-    private BigDecimal remaining;
+    private Double totalDonated;
+    private Double totalDistributed;
+    private Double remaining;
     private Long totalDonors;
     private Long totalBeneficiaries;
     private Long totalAssignments;
